requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.2.2
python-decouple==3.8
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.6.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-extensions==3.2.3
django-test-without-migrations==0.6
django-bootstrap4==24.3
djangorestframework==3.15.2
